requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
psycopg2>=2.9,<3.0  #
gunicorn>=20.0,<21.0 
python-dotenv>=0.19,<1.0
django-debug-toolbar>=3.2,<4.0 
whitenoise>=5.0,<6.0 
django-cors-headers==4.7.07
requests==2.32.3